requirements.txt
albumentations==1.2.1
h5py==3.6.0
matplotlib==3.5.1
numpy==1.21.5
opencv_python_headless==4.6.0.66
Pillow==9.2.0
scipy==1.7.3
torch==1.12.0
torchvision==0.13.0
